Breakout progress helper into its own module
